Split tests into dedicated files
